run: remove parallel tests capability for Windows/Mac

+++ run.py
@@ -43,6 +43,7 @@ from collections import namedtuple
 import yaml
 
 WIN32 = sys.platform == "win32"
+LINUX = sys.platform.startswith("linux")
 suricata_bin = "src\suricata.exe" if WIN32 else "./src/suricata"
 suricata_yaml = "suricata.yaml" if WIN32 else "./suricata.yaml"
 
@@ -877,8 +878,6 @@ def main():
 
     parser = argparse.ArgumentParser(description="Verification test runner.")
     parser.add_argument("-v", dest="verbose", action="store_true")
-    parser.add_argument("-j", type=int, default=min(8, mp.cpu_count()),
-                        help="Number of jobs to run")
     parser.add_argument("--force", dest="force", action="store_true",
                         help="Force running of skipped tests")
     parser.add_argument("--fail", action="store_true",
@@ -898,6 +897,9 @@ def main():
     parser.add_argument("--debug-failed", dest="debugfailed", action="store_true",
                         help="Prints debug output for failed tests")
     parser.add_argument("patterns", nargs="*", default=[])
+    if LINUX:
+        parser.add_argument("-j", type=int, default=min(8, mp.cpu_count()),
+                        help="Number of jobs to run")
     args = parser.parse_args()
 
     if args.self_test:
@@ -967,20 +969,28 @@ def main():
     # Sort alphabetically.
     tests.sort()
 
-    jobs = args.j
-    print("Number of concurrent jobs: %d" % jobs)
+    if LINUX:
+        jobs = args.j
+        print("Number of concurrent jobs: %d" % jobs)
 
-    pool = mp.Pool(jobs)
+        pool = mp.Pool(jobs)
     
-    try:
-        for dirpath in tests:
-            pool.apply_async(run_test, args=(dirpath, args, cwd, suricata_config))
-    except TerminatePoolError:
-        pool.terminate()
+        try:
+            for dirpath in tests:
+                pool.apply_async(run_test, args=(dirpath, args, cwd, suricata_config))
+        except TerminatePoolError:
+            pool.terminate()
 
-    pool.close()
-    pool.join()
-    
+        pool.close()
+        pool.join()
+
+    else:
+        try:
+            for dirpath in tests:
+                run_test(dirpath, args, cwd, suricata_config)
+        except TerminatePoolError:
+            sys.exit(1)
+
     passed = count_dict["passed"]
     failed = count_dict["failed"]
     skipped = count_dict["skipped"]
